Added correlation coefficient and changed splitting data
I changed the way i use the predict function, which now returns a vector of values. The splitting of the data into training and test set was not working correctly with the sample.split function, it did not return the correct amount of data in the test and training set. Therefore, i wrote new code that does work. All the queries that are not used were removed.